Steel lib project (lib export for other project, based on steel engine)
git-svn-id: svn://localhost/steel.new/trunk@209 99e027b2-380c-0410-a737-a5784e1a7bde

+++ steel/trunk/code/objects/game_obj.h
@@ -79,10 +79,10 @@ public:
 	
 	void addChildren(GameObj *obj);
 
-	int getPhysicChildrenCount(void)	{		return physicChildren.size();	}
+	int getPhysicChildrenCount(void)	{		return (int)physicChildren.size();	}
 	PhysicObject* getPhysicChildren(int i)	{		return physicChildren[i];	}
 
-	int getGraphChildrenCount(void)	{		return graphChildren.size();	}
+	int getGraphChildrenCount(void)	{		return (int)graphChildren.size();	}
 	GraphObject* getGraphChildren(int i)	{		return graphChildren[i];	}
 
 	GameObj *getParent() { return parent; }
